Rename commentGutterExtension to commentSelectionExtension
The file no longer implements a gutter - it provides the selection-based
tooltip. Rename to match its actual purpose.
